requirements.txt
Pillow>=10.0.0
numpy>=1.24.0
Flask>=3.0.0


